repo,server,models: Move AutomergeActions type to desci-models, support removing dpid
