+++ desci-models/src/index.ts
@@ -3,3 +3,4 @@ export * from './constants';
 export * from './transformers/index';
 export * from './trees/treeTools';
 export * from './trees/treeTypes';
+export * from './automerge';
