+++ desci-repo/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@desci-labs/desci-repo",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "",
   "type": "module",
   "engines": {
