+++ desci-server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@desci-labs/desci-server",
-  "version": "0.3.3",
+  "version": "0.3.4",
   "description": "Manage DeSci Nodes state: users, nodes orchestration, decentralized storage, blockchain resolvers",
   "engines": {
     "node": ">=16.14"
