+++ desci-server/src/services/repoService.ts
@@ -1,11 +1,11 @@
 import { DocumentId } from '@automerge/automerge-repo';
-import { ResearchObjectV1 } from '@desci-labs/desci-models';
+import { ResearchObjectV1, ManifestActions } from '@desci-labs/desci-models';
 import axios, { AxiosInstance } from 'axios';
 
 import { logger as parentLogger } from '../logger.js';
 import { ResearchObjectDocument } from '../types/documents.js';
 
-import { ManifestActions, NodeUuid } from './manifestRepo.js';
+import { NodeUuid } from './manifestRepo.js';
 
 const logger = parentLogger.child({ module: 'Repo Service' });
